feat: add approtium model

+++ src/config.py
@@ -1,6 +1,6 @@
 import os
 
-DEFAULT_PRECISION = "fp32"
+DEFAULT_PRECISION = "fp16"
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 STATIC_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "main/static"))
